test(jsx): keep setup wrapper tests under source limit
